Delete code.
